refactor: :recycle: Implemented useEffect on Categories
